+++ lib/ex_data_sketch/codec.ex
@@ -31,6 +31,7 @@ defmodule ExDataSketch.Codec do
   - 4: KLL (Quantiles)
   - 5: DDSketch (Quantiles)
   - 6: FrequentItems (SpaceSaving)
+  - 7: Bloom
 
   ## Versioning
 
@@ -49,8 +50,9 @@ defmodule ExDataSketch.Codec do
   @sketch_id_kll 4
   @sketch_id_ddsketch 5
   @sketch_id_fi 6
+  @sketch_id_bloom 7
 
-  @type sketch_id :: 1 | 2 | 3 | 4 | 5 | 6
+  @type sketch_id :: 1 | 2 | 3 | 4 | 5 | 6 | 7
   @type decoded :: %{
           version: pos_integer(),
           sketch_id: sketch_id(),
@@ -155,11 +157,23 @@ defmodule ExDataSketch.Codec do
   def sketch_id_fi, do: @sketch_id_fi
 
   @doc """
+  Returns the sketch ID constant for Bloom.
+
+  ## Examples
+
+      iex> ExDataSketch.Codec.sketch_id_bloom()
+      7
+
+  """
+  @spec sketch_id_bloom() :: sketch_id()
+  def sketch_id_bloom, do: @sketch_id_bloom
+
+  @doc """
   Encodes sketch data into the EXSK binary format.
 
   ## Parameters
 
-  - `sketch_id` - sketch type identifier (1=HLL, 2=CMS, 3=Theta, 4=KLL, 5=DDSketch, 6=FrequentItems)
+  - `sketch_id` - sketch type identifier (1=HLL, 2=CMS, 3=Theta, 4=KLL, 5=DDSketch, 6=FrequentItems, 7=Bloom)
   - `version` - format version (use `Codec.version/0` for current)
   - `params_bin` - binary-encoded sketch parameters
   - `state_bin` - raw sketch state binary