+++ test/ex_data_sketch_backend_test.exs
@@ -36,6 +36,12 @@ defmodule ExDataSketch.BackendTest.StubBackend do
   def ddsketch_count(_s, _o), do: 0
   def ddsketch_min(_s, _o), do: nil
   def ddsketch_max(_s, _o), do: nil
+  def bloom_new(_opts), do: <<>>
+  def bloom_put(s, _h, _o), do: s
+  def bloom_put_many(s, _h, _o), do: s
+  def bloom_member?(_s, _h, _o), do: false
+  def bloom_merge(s, _b, _o), do: s
+  def bloom_count(_s, _o), do: 0
   def fi_new(_opts), do: <<>>
   def fi_update(s, _ib, _o), do: s
   def fi_update_many(s, _items, _o), do: s